@@ -1,22 +1,20 @@
 import FileManager from './FileManager';
-
-const IPFS = require('ipfs-mini');
-const fs = require('fs');
+import DeployInfo from './DeployInfo';
 
 class IPFSFileManager implements FileManager {
   private ipfs : any;
 
-  constructor() {
-    this.ipfs = new IPFS({ host: 'ipfs.infura.io', port: 5001, protocol: 'https' });
+  constructor(ipfsObj : any) {
+    this.ipfs = ipfsObj;
   }
 
   /**
   * @returns The IPFS CID related to the uploaded JSON
-  * @param buffer The buffer you want to save to IPFS
+  * @param buffer The deployment info you want to save to IPFS
   */
-  public async save(buffer: Buffer) : Promise<string> {
+  public async save(deployInfo : DeployInfo) : Promise<string> {
     return new Promise((resolve, reject) => {
-      this.ipfs.add(buffer.toString('hex'))
+      this.ipfs.addJSON(deployInfo)
         .then(resolve)
         .catch((error : Error) => {
           reject(new Error(`It seems that there are some problems with IPFS, error: ${error}`));
@@ -28,10 +26,10 @@ class IPFSFileManager implements FileManager {
   * @returns The JSON object corresponding to the given CID
   * @param {*} hash The IPFS CID has you want to get the JSON
   */
-  public async get(cid : string) : Promise<Buffer> {
+  public async get(cid : string) : Promise<DeployInfo> {
     return new Promise((resolve, reject) => {
-      this.ipfs.cat(cid)
-        .then((result : string) => resolve(Buffer.from(result, 'hex')))
+      this.ipfs.catJSON(cid)
+        .then((result : DeployInfo) => resolve(result))
         .catch((error : Error) => {
           reject(new Error(`It seems that there are some problems with IPFS, error: ${error}`));
         });
