@@ -8,18 +8,31 @@ const fs = require('fs');
 class SignupCommand extends Command {
   command = 'signup [save]';
 
-  description = 'create a new account';
+  description = 'Description:\n_\b  Create a new account';
 
+  /**
+   * @method exec
+   * @param yargs: arguments nedded for the command
+   * @description the command creates a new ethereum wallet and return a string
+   *  with all its information. If the save flag is indicated, the wallet credentials
+   *  will be saved in a file.
+   */
   async exec(args: any) : Promise<any> {
     const wallet : Wallet = this.session.signup();
 
+    // check if the user request to save the credentials
     if (args.save) {
-      fs.writeFile('./credential.txt', `Address: ${wallet.address} \nPrivate Key: ${wallet.privateKey} \nMnemonic phrase: ${wallet.mnemonic}`);
+      fs.writeFileSync('./credentials.txt', `Address: ${wallet.address} \nPrivate Key: ${wallet.privateKey} \nMnemonic phrase: ${wallet.mnemonic}`);
     }
 
+    // return wallet credentials
     return `Address: ${wallet.address} \nPrivate Key: ${wallet.privateKey} \nMnemonic phrase: ${wallet.mnemonic}`;
   }
 
+  /**
+   * Descriptor of the command
+   * @param yargs: object used to define the command params
+   */
   builder(yargs : Argv) : any {
     return yargs.option('save', {
       describe: 'Decide if save credentials in file',
