@@ -1,4 +1,6 @@
+import DeployInfo from './DeployInfo';
+
 export default interface FileManager {
-  save(data : Buffer) : Promise<string>;
-  get(CID : string) : Promise<Buffer>;
+  save(data : DeployInfo) : Promise<string>;
+  get(CID : string) : Promise<DeployInfo>;
 }
